requirements.txt
python-dotenv~=1.0.1
netmiko~=4.5.0
requests~=2.32.3